acq: move acquisition manager code to separate acqmng module
It looked fine when using it (eg, "acq.acquire()"), but as the code
inside it needs objects defined in sub-modules, that meant the
sub-modules could import this parent modules due to cyclic dependency.

We had managed so far, but it's "cleaner" to have the main code in a
separate module.
