+++ plugins/ar_spectral.py
@@ -47,8 +47,8 @@ from concurrent.futures._base import CancelledError, CANCELLED, FINISHED, RUNNIN
 import logging
 import math
 import numpy
-from odemis import dataio, model, acq
-from odemis.acq import stream, drift
+from odemis import dataio, model
+from odemis.acq import stream, drift, acqmng
 from odemis.acq.stream import UNDEFINED_ROI
 import odemis.gui
 from odemis.gui.conf import get_acqui_conf
@@ -996,7 +996,7 @@ class ARspectral(Plugin):
         exporter = dataio.find_fittest_converter(fn)
 
         try:
-            f = acq.acquire(strs, self.main_app.main_data.settings_obs)
+            f = acqmng.acquire(strs, self.main_app.main_data.settings_obs)
             dlg.showProgress(f)
             das, e = f.result()  # blocks until all the acquisitions are finished
         except CancelledError:
